+++ R/translate-sql-base.r
@@ -158,7 +158,7 @@ base_win <- sql_translator(
   min   = win_recycled("min"),
   max   = win_recycled("max"),
   n     = function() {
-    over(sql("COUNT(*)"), partition_group(), frame = c(-Inf, Inf))
+    over(sql("COUNT(*)"), partition_group())
   },
 
   # Cumulative function are like recycled aggregates except that R names